requirements.txt
gspread
Pillow
oauth2client
boto3
facebook-sdk
